requirements.txt
aioflask==0.4.0
APScheduler==3.9.1
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.1.1
greenlet==1.1.2
greenletio==0.9.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
toml==0.10.2
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.0.3
zipp==3.8.0
